Change header file extension
