+++ src/Data.hpp
@@ -21,7 +21,7 @@
   */
 #pragma once
 
-#include "Weather.h"
+#include "Weather.hpp"
 #include <nvs.h>
