+++ src/SHT30.hpp
@@ -15,40 +15,21 @@
    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 /**
-  * @file Battery.h
+  * @file SHT30.h
   * 
-  * Helperfunctions for reading the battery value.
+  * Helper function to read the SHT30 environment data.
   */
 #pragma once
+#include "Data.hpp"
 
-#include "Data.h"
-
-/**
-  * Read the battery voltage
-  */
-bool GetBatteryValues(MyData &myData)
+/* Read the SHT30 environment chip data */
+bool GetSHT30Values(MyData &myData)
 {
-   uint32_t vol = M5.getBatteryVoltage();
-
-   if (vol < 3300) {
-      vol = 3300;
-   } else if (vol > 4350) {
-      vol = 4350;
+   M5.SHT30.UpdateData();
+   if(M5.SHT30.GetError() == 0) {
+      myData.sht30Temperatur = (int) M5.SHT30.GetTemperature();
+      myData.sht30Humidity   = (int) M5.SHT30.GetRelHumidity();
+      return true;
    }
-  
-   float battery = (float)(vol - 3300) / (float)(4350 - 3300);
-
-   myData.batteryVolt = vol / 1000.0f;
-   Serial.println("batteryVolt: " + String(myData.batteryVolt));
-   
-   if (battery <= 0.01) {
-      battery = 0.01;
-   }
-   if (battery > 1) {
-      battery = 1;
-   }
-   myData.batteryCapacity = (int) (battery * 100);
-   Serial.println("batteryCapacity: " + String(myData.batteryCapacity));
-   
-   return true;
+   return false;
 }
